+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVMB8PHGR849N3Z92WPGDJRA"
+        "01JVP48W96H0M8MW286EK1YD64"
     )
     r"""The id of the resource"""