+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-19T13:20:49.142Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-20T05:57:03.889Z")
     r"""The date and time the resource was last updated"""