+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVMB8PHHS0WNYG0NXPXD7NXG"
+        "01JVP48W97X7SJR5AQQ4PTNCBM"
     )
     r"""The id of the resource"""