+++ packages/orq-rc/RELEASES.md
@@ -1289,3 +1289,13 @@ Based on:
 - [python v3.6.0-rc.14] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.14] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.14 - packages/orq-rc
+
+## 2025-05-20 05:58:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.548.1 (2.605.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.15] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.15] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.15 - packages/orq-rc